Implement DELETE user to heroku server

+++ app/src/main/java/my/com/icheckin/icheckin_android/view/ModifyAccountActivity.kt
@@ -2,19 +2,30 @@ package my.com.icheckin.icheckin_android.view
 
 import android.os.Bundle
 import android.support.v4.content.ContextCompat
+import android.support.v7.app.AlertDialog
 import android.support.v7.app.AppCompatActivity
+import android.view.WindowManager
 import com.pawegio.kandroid.textWatcher
 import com.pawegio.kandroid.toast
 import kotlinx.android.synthetic.main.activity_modify_account.*
+import kotlinx.coroutines.experimental.android.UI
+import kotlinx.coroutines.experimental.async
+import kotlinx.coroutines.experimental.launch
 import my.com.icheckin.icheckin_android.R
+import my.com.icheckin.icheckin_android.controller.Izone
 import my.com.icheckin.icheckin_android.model.entity.Credential
 import my.com.icheckin.icheckin_android.utils.storage.AppDatabase
+import my.com.icheckin.icheckin_android.utils.view.CustomProgressDialog
 
 class ModifyAccountActivity : AppCompatActivity() {
+
+    private lateinit var progressDialog: CustomProgressDialog
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_modify_account)
+
+        progressDialog = CustomProgressDialog(this, "Deleting...")
 
         val db = AppDatabase.getDatabase(applicationContext)
         val credential = intent.getSerializableExtra("credential") as Credential
@@ -36,9 +47,31 @@ class ModifyAccountActivity : AppCompatActivity() {
         }
 
         imageButton_Delete.setOnClickListener {
-            db.credentialDao().delete(credential)
-            toast("${credential.username} deleted")
-            finish()
+            AlertDialog.Builder(this)
+                    .setTitle("Delete credential")
+                    .setMessage("Do you want to remove credential from server?")
+                    .setPositiveButton("Remove from server") { _, _ ->
+                        progressDialog.show()
+                        window.addFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND)
+                        async {
+                            Izone.delete(credential)
+                            launch(UI) {
+                                progressDialog.hide()
+                                window.clearFlags(WindowManager.LayoutParams.FLAG_DIM_BEHIND)
+                                db.credentialDao().delete(credential)
+                                toast("${credential.username} deleted")
+                                finish()
+                            }
+                        }
+                    }
+                    .setNegativeButton("This device only") { _, _ ->
+                        db.credentialDao().delete(credential)
+                        toast("${credential.username} deleted")
+                        finish()
+                    }
+                    .setCancelable(false)
+                    .create()
+                    .show()
         }
 
         imageButton_Close.setOnClickListener {
